fixed gui listings after template deletion
The issue was that the owner was not set correctly.

This change will make the _local_company available as this->owner within
the TbtrGui - this means we don't have to pass around the owner as a
parameter to each method.

Don't know if setting the this->owner has any side effects, since it is
explicitly set to INVALID_OWNER by FinishInitNested(...).