removed ChainContainsVehicle(...)
This was only called from Find*InDepot(). It's not needed since it
should only prevent the case that when we are looking for a vehicle in
the depot we don't accidentally consider one from the incoming chain.

But this doesn't happen because the incoming chain is always checked
before the depot. If the depot is even checked at all.