inherit from Window instead of BuildVehicleWindow
It woupldn't compile the other way. The problem is that the window
descriptor has to be passed to the parent class. But the
BuildVehicleWindow has no header files and cannot be included. One
use its ctor as parent ctor if one only forward-declared the class
either. And the Window() ctor can only be used when directly inheriting
from Window.