fixed composition of articulated engines
Problem:
When cloning from a train, a single articulated engine would be cloned
as a single vehicle part, missing the articulated part.
BUT, when cloning an articulated engine with a wagon, both artic parts
of the engine would be cloned.

This has been made consistent now, so that every part of an artic engine
will be cloned into the template. This should be the preferred behaviour
anyway since during drawing and template replacement, the artic parts
are skipped over anyways.