fixed crash when creating template chain
The problem appeared when a template was created by cloning from a train
or when loading a savegame which had stored templates.

The initialization of the sprite size can only be done when:
a) the template has a well-defined engine_type
b) the Engine::Get() call can be used inside the GetTrainSpriteSize()
   function

At least the 2nd point cannot be guaranteed during the initialization
phase, e.g. the game would always crash during savegame load, even when
separating the sprite size initialization into a different function and
making sure that it was only called when an engine_type had been set.

Now the caching of the sprite size is only done when it's needed (during
drawing) and if it has not been done already. This makes it also safe to
use the default ctor without the need of explicitly calling a separate
initialization routine.