restore original chain after failed tbtr
A template replacement may fail due to a bad template composition, e.g.
the wrong wagons for a given locomotive. This cannot be detected
beforehand because the CMD_MOVE_RAIL_VEHICLE will always return
success==false when testing the tbtr command, i.e. during flags==DC_NONE.

So, in case that a train could not be composed as the template
specifies, the original chain will be backed up (by vehicle id only) so
that it can be restored later. This is possible because after the main
train-composition loop all vehicles still exist and no orders etc have
been changed until that point.